Source /etc/profile before running tool tests
Some packages modify $PATH variable to include ~/.dotnet/tools by adding
a file to /etc/profile.d/. This path is re-sourced on every login. But
not otherwise. So the changes to $PATH only happen after a user logs out
and back in after installing the dotnet package.

Make the tests simulate that by re-sourcing /etc/profile so the profile
settings are re-loaded.