extract-bundle-dir is only valid for 3.1
This feature was introduced with .NET Core 3.0 (now EOL). In .NET 5.0,
the single standalone file is not extracted to a on-disk location. This
test is not needed there. Because the bundle extraction directory is
empty, the test also results in a false positive.

See https://github.com/dotnet/runtime/issues/36590 for more information.