Enable auto-fixing the TFM in use-current-runtime
The test runner looks for project files named the same as the directory
and updates the TargetFramework element in them. This file was named
differently, so it wasn't being identified by the test runner and the
TFM wasn't being updated.

The result was that testing against a .NET 7 SDK would still build and
run the application against .NET 6.