Handle additional changes in dotnet dump output
The new, in-development, versions of dotnet dump have made several
changes to the otuput of the dotnet-analyze command:

- The output format of `dumpruntimetypes` doesn't include leading 0's anymore:

         Address           Domain               MT Type Name
    7ed6c0800580     55fb3bed7b10     7f16a5b11038 System.String

- The `threadpool` command now displays output in this format:

     CPU utilization: 0%
       Workers Total: 3
     Workers Running: 0
	Workers Idle: 3
    Worker Min Limit: 20
    Worker Max Limit: 32767

- The `traverseheap` command no longer accepts a `-verify` option.

Make the test more flexible and accept both the new and old styles of
output as much as possible.

Also fix bad uses of `wc -l dump.out`: that prints the number of lines
and the filename. If we just want the number of lines, we need to
extract it from the output.